finai-hackops/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
passlib[bcrypt]
python-multipart
python-jose[cryptography]
pydantic[email]
google-auth
pandas
scikit-learn
joblib
